This is mostly a bugfix release

* Fix 17 bugs, most of them securety related
* Fix a number of issues uncovered by Coverity
* Add meson option to enable unstable APIs and add the info to pkg-config
* vtest: Stabilize protocol version 3

* Virgl:
  - Refactor virgl command stream decoding to check parameters early
  - Improve format support also for video
  - Add support for VIRGL_CCMD_CLEAR_SURFACE
  - Fall back to OpenGL when GBM mappinmg fails

* Vkr/Venus:
  - VK_EXT_graphics_pipeline_library
  - add tracepoints for profiling and debugging
  - improve validation
  - Implement VIRGL_RESOURCE_FD_SHM
  - Introduce a new capset to require guest-based blob allocations

+++ tests/test_virgl_resource.c
@@ -36,7 +36,7 @@
 #include "testvirgl.h"
 
 #include "pipe/p_defines.h"
-#include "pipe/p_format.h"
+#include "util/u_formats.h"
 
 #ifndef ARRAY_SIZE
 #  define ARRAY_SIZE(arr) (sizeof(arr) / sizeof((arr)[0]))
@@ -48,44 +48,44 @@ struct res_test {
 };
 
 #define TEST(thandle, ttarget, tformat, tbind, twidth, theight, tdepth, tarray_size, tnr_samples, tretval) \
-  { .args = { .handle = (thandle),					\
-	      .target = (ttarget),					\
-	      .format = (tformat),					\
-	      .bind = (tbind),						\
-	      .width = (twidth),					\
-	      .height = (theight),					\
-	      .depth = (tdepth),					\
-	      .array_size = (tarray_size),				\
-	      .last_level = 0,						\
-	      .nr_samples = (tnr_samples),				\
-	      .flags = 0 },					\
+  { .args = { .handle = (thandle),                                      \
+              .target = (ttarget),                                      \
+              .format = (tformat),                                      \
+              .bind = (tbind),                                          \
+              .width = (twidth),                                        \
+              .height = (theight),                                      \
+              .depth = (tdepth),                                        \
+              .array_size = (tarray_size),                              \
+              .last_level = 0,                                          \
+              .nr_samples = (tnr_samples),                              \
+              .flags = 0 },                                             \
       .retval = (tretval)}
 
 #define TEST_MIP(thandle, ttarget, tformat, tbind, twidth, theight, tdepth, tarray_size, tnr_samples, tlast_level, tretval) \
-  { .args = { .handle = (thandle),					\
-	      .target = (ttarget),					\
-	      .format = (tformat),					\
-	      .bind = (tbind),						\
-	      .width = (twidth),					\
-	      .height = (theight),					\
-	      .depth = (tdepth),					\
-	      .array_size = (tarray_size),				\
-	      .last_level = (tlast_level),				\
-	      .nr_samples = (tnr_samples),				\
-	      .flags = 0 },					\
+  { .args = { .handle = (thandle),                                      \
+              .target = (ttarget),                                      \
+              .format = (tformat),                                      \
+              .bind = (tbind),                                          \
+              .width = (twidth),                                        \
+              .height = (theight),                                      \
+              .depth = (tdepth),                                        \
+              .array_size = (tarray_size),                              \
+              .last_level = (tlast_level),                              \
+              .nr_samples = (tnr_samples),                              \
+              .flags = 0 },                                             \
       .retval = (tretval)}
 
 #define TEST_F(thandle, ttarget, tformat, tbind, twidth, theight, tdepth, tarray_size, tnr_samples, tflags, tretval) \
-  { .args = { .handle = (thandle),					\
-	      .target = (ttarget),					\
-	      .format = (tformat),					\
-	      .bind = (tbind),						\
-	      .width = (twidth),					\
-	      .height = (theight),					\
-	      .depth = (tdepth),					\
-	      .array_size = (tarray_size),				\
-	      .nr_samples = (tnr_samples),				\
-	      .flags = (tflags) },					\
+  { .args = { .handle = (thandle),                                      \
+              .target = (ttarget),                                      \
+              .format = (tformat),                                      \
+              .bind = (tbind),                                          \
+              .width = (twidth),                                        \
+              .height = (theight),                                      \
+              .depth = (tdepth),                                        \
+              .array_size = (tarray_size),                              \
+              .nr_samples = (tnr_samples),                              \
+              .flags = (tflags) },                                      \
       .retval = (tretval)}
 
 
@@ -299,7 +299,7 @@ START_TEST(private_ptr)
   void *init_priv = (void*)0xabab;
   virgl_renderer_resource_set_priv(1, init_priv);
   void *priv = virgl_renderer_resource_get_priv(1);
-  ck_assert_int_eq((unsigned long)priv, 0xabab);
+  ck_assert_int_eq((uintptr_t)priv, 0xabab);
   testvirgl_fini_single_ctx();
 }
 END_TEST
